+++ src/DNIC.Erechtheion.Core/ISeedDataInitiator.cs
@@ -1,0 +1,7 @@
+﻿namespace DNIC.Erechtheion.Core
+{
+	public interface ISeedDataInitiator
+	{
+		void EnsureSeedData();
+	}
+}
