--- src/DNIC.Erechtheion.Domain/IRepositorySeedData.cs
@@ -1,11 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace DNIC.Erechtheion.Domain
-{
-	public interface IRepositorySeedData
-	{
-		void EnsureSeedData();
-	}
-}
