--- src/DNIC.Erechtheion.EntityFrameworkCore/ErechtheionUser.cs
@@ -1,13 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-using Microsoft.AspNetCore.Identity;
-
-namespace DNIC.Erechtheion.EntityFrameworkCore
-{
-	// Add profile data for application users by adding properties to the ApplicationUser class
-	public class ErechtheionUser : IdentityUser
-	{
-	}
-}
