+++ src/DNIC.Erechtheion.Identity.EntityFrameworkCore/ErechtheionUser.cs
@@ -1,0 +1,9 @@
+﻿using Microsoft.AspNetCore.Identity;
+
+namespace DNIC.Erechtheion.Identity.EntityFrameworkCore
+{
+	// Add profile data for application users by adding properties to the ApplicationUser class
+	public class ErechtheionUser : IdentityUser
+	{
+	}
+}
