+++ src/DNIC.Erechtheion/DNIC.Erechtheion.csproj
@@ -27,5 +27,6 @@
     <ProjectReference Include="..\DNIC.Erechtheion.Application\DNIC.Erechtheion.Application.csproj" />
     <ProjectReference Include="..\DNIC.Erechtheion.Core\DNIC.Erechtheion.Core.csproj" />
     <ProjectReference Include="..\DNIC.Erechtheion.EntityFrameworkCore\DNIC.Erechtheion.EntityFrameworkCore.csproj" />
+    <ProjectReference Include="..\DNIC.Erechtheion.Identity.EntityFrameworkCore\DNIC.Erechtheion.Identity.EntityFrameworkCore.csproj" />
   </ItemGroup>
 </Project>
