+++ src/DNIC.Erechtheion/DependencyInjectionConfig.cs
@@ -1,11 +1,13 @@
 ﻿using DNIC.Erechtheion.Application;
 using DNIC.Erechtheion.Application.Service;
+using DNIC.Erechtheion.Core;
 using DNIC.Erechtheion.Core.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using DNIC.Erechtheion.EntityFrameworkCore;
 using DNIC.Erechtheion.Domain;
 using DNIC.Erechtheion.Domain.Repositories;
 using DNIC.Erechtheion.EntityFrameworkCore.Repositories;
+using DNIC.Erechtheion.Identity.EntityFrameworkCore;
 
 namespace DNIC.Erechtheion
 {
@@ -14,7 +16,8 @@ namespace DNIC.Erechtheion
 		public static void AddErechtheion(this IServiceCollection services)
 		{
 			services.AddSingleton<IErechtheionConfiguration, ErechtheionConfiguration>();
-			services.AddSingleton<IRepositorySeedData, RepositorySeedData>();
+			services.AddSingleton<ISeedDataInitiator, SeedDataInitiator>();
+			services.AddSingleton<ISeedDataInitiator, IdentitySeedDataInitiator>();
 			services.AddScoped<ITopicRepository, TopicRepository>();
 			services.AddScoped<ITopicApplicationService, TopicApplicationService>();
 		}
