+++ src/DNIC.Erechtheion/Views/Account/Login.cshtml
@@ -1,5 +1,7 @@
 ﻿@using System.Collections.Generic
 @using System.Linq
+@using DNIC.Erechtheion.Identity
+@using DNIC.Erechtheion.Identity.EntityFrameworkCore
 @using Microsoft.AspNetCore.Http
 @using Microsoft.AspNetCore.Http.Authentication
 @model LoginViewModel
