+++ src/DNIC.Erechtheion/Views/Manage/_ManageNav.cshtml
@@ -1,4 +1,6 @@
-﻿@using DNIC.Erechtheion.Views.Manage
+﻿@using DNIC.Erechtheion.Identity
+@using DNIC.Erechtheion.Identity.EntityFrameworkCore
+@using DNIC.Erechtheion.Views.Manage
 @inject SignInManager<ErechtheionUser> SignInManager
 @{
     var hasExternalLogins = (await SignInManager.GetExternalAuthenticationSchemesAsync()).Any();
@@ -13,4 +15,3 @@
     }
     <li class="@ManageNavPages.TwoFactorAuthenticationNavClass(ViewContext)"><a asp-action="TwoFactorAuthentication">Two-factor authentication</a></li>
 </ul>
-
